# ENH: Speed up trim_zeros