--- a/doc/release/upcoming_changes/16911.deprecation.rst
+++ b/doc/release/upcoming_changes/16911.deprecation.rst
@@ -1,0 +1,7 @@
+``trim_zeros`` now requires a 1D array compatible with ``ndarray.astype(bool)``
+-------------------------------------------------------------------------------
+The ``trim_zeros`` function will, in the future, require an array with the
+following two properties:
+
+* It must be 1D.
+* It must be convertable into a boolean array.