--- a/numpy/lib/tests/test_function_base.py
+++ b/numpy/lib/tests/test_function_base.py
@@ -1169,14 +1169,10 @@ class TestTrimZeros:
     a = np.array([0, 0, 1, 0, 2, 3, 4, 0])
     b = a.astype(float)
     c = a.astype(complex)
-    d = a.astype([('a', int, (2, 3)), ('b', float), ('c', bool)])
-
-    e = np.array([None, [], 1, False, 'b', 3.0, range(4), b''], dtype=object)
-    f = np.array(['', '', 'a', '', 'b', 'c', 'd', ''])
-    g = f.astype(bytes)
+    d = np.array([None, [], 1, False, 'b', 3.0, range(4), b''], dtype=object)
 
     def values(self):
-        attr_names = ('a', 'b', 'c', 'd', 'e', 'f', 'g')
+        attr_names = ('a', 'b', 'c', 'd')
         return (getattr(self, name) for name in attr_names)
 
     def test_basic(self):