--- a/numpy/lib/tests/test_function_base.py
+++ b/numpy/lib/tests/test_function_base.py
@@ -1208,22 +1208,6 @@ class TestTrimZeros:
         res = trim_zeros(arr)
         assert_array_equal(arr, res)
 
-    def test_deprecation(self):
-        arr = np.random.rand(10, 10).tolist()
-
-        with warnings.catch_warnings():
-            warnings.simplefilter('error', DeprecationWarning)
-            try:
-                trim_zeros(arr)
-            except Exception as ex:
-                assert isinstance(ex, DeprecationWarning)
-                assert isinstance(ex.__cause__, ValueError)
-            else:
-                raise AssertionError('Failed to raise an exception')
-
-        out = nfb._trim_zeros_old(arr)
-        assert_allclose(out, arr)
-
 
 class TestExtins: