Closes https://github.com/numpy/numpy/issues/16783.

As was noted in aforementioned issue the current `trim_zeros()` implementation, 
as of 1.19.0, is very slow with plenty of room for further optimization. 
This pull request addresses the previous optimization issue. 

Before
-------
``` python
In [1]: import numpy as np

In [2]: a = np.hstack([
   ...:     np.zeros(100_000),
   ...:     np.random.uniform(size=100_000),
   ...:     np.zeros(100_000),
   ...: ])

In [3]: np.__version__
Out[3]: '1.19.0'

In [4]: %timeit np.trim_zeros(a)
45.8 ms ± 6.38 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After
-----
``` python
In [3]: np.__version__
Out[3]: '1.20.0.dev0+2823c98'

In [4]: %timeit np.trim_zeros(a)
303 µs ± 15 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

To summarize the new implementation: the passed array is now first converted a boolean array,
after which `argmax()` is used for identifying the first and/or last non-zero element.

~~A side effect of the new approach is that it will trim any leading and/or trailing elements
which evaluates to ``False``; not just ``0``.~~

Lastly, a new benchmark has been added as was requested in https://github.com/numpy/numpy/issues/16783#issuecomment-655856960.