@BvB93 2020-07-20
> I hadn't realized that `trim_zeros` wasn't written for arrays at all.

Yeah, it's most definitely a bit of an oddball.

> I worry though that we can't change it, else code like `trim_zeros([[1], 0])` will stop working.

Object arrays should work fine as long as they can be either converted into boolean arrays (such as is the case with your example).
``` python
In [1]: import numpy as np                                                                                                                             

In [2]: np.array([[1], 0], dtype=object).astype(bool)                                                                                                  
Out[2]: array([ True, False])
```

The only exception I can think of is a (ragged) object array consisting of other arrays, 
though these didn't work with the old implementation either is it involved array to scalar comparisons.

``` python
In [1]: import numpy as np

In [2]: a = np.empty(2, dtype=object)

In [3]: a[0] = a[1] = np.random.rand(5,)

In [4]: a.astype(bool)
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
    ...
ValueError: setting an array element with a sequence.```