@BvB93 2020-07-20
Circleci failure seems to be caused by the unrelated `histogram2d()` documentation.