@eric-wieser 2020-07-20
> Object arrays should work fine as long as they can be either converted into boolean arrays (such as is the case with your example).

Right, but your example code only works because you explicitly added `dtype=object`. From what I can tell, this fails / emits a deprecation warning with your code;
```python
>>> np.trim_zeros([[1], 0])
[[1]]
```
and this fails outright:
```python
>>> np.trim_zeros([[1]])
[[1]]
```