@seberg 2020-07-20
As silly as `trim_zeros` is, I think the reason it uses a for loop, is probably speed to begin with: It is probably *much* faster as is if you only trim a handfull of zeros (or even none).  However, its plausible that you can argue it does not matter:  Its like having a fast-path for `np.all()` when the first element is already `False`:  It saves an arbitrary amount of time, but if you do any _other_ operation on the array later, the speedup is probably dwarfed by that.