@BvB93 2020-07-20
> As silly as `trim_zeros` is, I think the reason it uses a for loop, is probably speed to begin with: It is probably much faster ~~as is~~ if you only trim a handfull of zeros (or even none). 

Indeed, the example execution times I provided are for a pretty large array; 
the difference will be much smaller for when the input array is smaller.

Fortunetly, even arrays as small as `np.ones(1)` show a minor improvements in execution time (~4 µs (new) vs ~4.5 µs (old)).