@BvB93 2020-07-20
> Fortunetly, even arrays as small as `np.ones(1)` show a minor improvements in execution time (~4 µs (new) vs ~4.5 µs (old)).

On second thought, I suspect this is moreso related to the old implementation's lack of `enumerate()` and the `if i == 0.:` rather than `if i:` comparisons