@BvB93 2020-07-20
> and this fails outright:
> ``` python
> >>> np.trim_zeros([[1]])
> [[1]]
> ```

After giving it a bit more thought there might be a very simple way of dealing with this issue: 
simply return the passed `filt` if it cannot be coerced into a 1D array (instead of raising a `ValueError`).

While I feel raising an error would generally speaking be more appropriate, 
this alternative solution is definitely safer as far as backwards compatibility is concerned.