@eric-wieser 2020-07-20
> Maybe in combination with a deprecation warning of some sort?

Perhaps the thing to do is something like:
```
if isinstance(arr, np.ndarray):
    # your optimal code
else:
   # emit deprecation warning
   # old code
```

Another more cynical view would be just deprecate this function entirely as it's trivial slow and unlike most of numpy, and work out a better name / submodule / PyPI package for the fast `trim_zeros`.