@BvB93 2020-07-20
> Perhaps the thing to do is something like:
> ```python
> ...
> ```

I just added something like this in https://github.com/numpy/numpy/pull/16911/commits/a6f9d296e79c8a1de65bfd18a17bdb7b1ed427c7.
It will now automatically fall back to the old implementation if an exception is encountered (where the same exception may or may not be raised again, depending on its nature.

> Another more cynical view would be just deprecate this function entirely as it's trivial slow and unlike most of numpy, and work out a better name / submodule / PyPI package for the fast `trim_zeros`.

If the option was just between leaving it in and removing it then I would agree with this view, 
as in its current state It's IMO not really something that belongs in numpy's public API.
However as `trim_zeros()` still seems salvageable, I'm personally leaning more towards leaving it in.