@mattip 2020-07-31
New deprecations require
-  [ ] a date/version comment like [this one](https://github.com/numpy/numpy/blob/master/numpy/core/_methods.py#L96),
- [ ] tests that the warning works in [test/test_deprecations](https://github.com/numpy/numpy/blob/master/numpy/core/tests/test_deprecations.py) and
- [ ] a [release note](https://github.com/numpy/numpy/tree/master/doc/release/upcoming_changes).