@BvB93 2020-07-31
Thanks @mattip! 
Your comments have, respectively, been addressed in https://github.com/numpy/numpy/pull/16911/commits/853b83a6dee09d190a0ee9caa099a5b8a1f8cd94, https://github.com/numpy/numpy/pull/16911/commits/65f5634808a925837c2ec41116a0b12ed162f510 and https://github.com/numpy/numpy/pull/16911/commits/d531cd7f735a47039a9ffc88493c8837c8d2a2fd.

As a side note, the branch has been rebased to get rid of a merge conflict.