@BvB93 2020-08-03
Still failing due to the ``setuptools`` issue.
Rebasing in order to apply https://github.com/numpy/numpy/pull/16993.