@eric-wieser 2020-08-04
I think this breaks `np.trim_zeros([2**64, 0])`. It should probably either mention this in the release notes, or we should emit a deprecation warning on all non-array inputs.