@mattip 2020-08-04
Both before and after this PR I get a list with a single item `[18446744073709551616]` on linux64, python 3.8. What am I missing?