@BvB93 2020-08-04
> I think this breaks `np.trim_zeros([2**64, 0])`.

Overflow shouldn't be an issue actually, as `np.array()` will automatically default to an alternative dtype if an integer is too large. Note that all below-mentioned data types work just fine with `ndarray.astype(bool)`.

``` python
In [1]: import numpy as np                                                                                             

In [2]: np.asanyarray(2**62)                                                                                                
Out[2]: array(4611686018427387904)

In [3]: np.asanyarray(2**63)                                                                                                
Out[3]: array(9223372036854775808, dtype=uint64)

In [4]: np.asanyarray(2**64)                                                                                                
Out[4]: array(18446744073709551616, dtype=object)
```