@mhvk 2020-08-11
Note that this breaks astropy [1] as we can compare Quantities to 0 but cannot turn them into bool (not clear that '0 m' should be `False`, after all). While it isn't really a big deal since we can override with `__array_function__`, I wondered if it would be OK to have a follow-up PR to change the `astype(bool)` to comparing with 0 -- it not only is closer to what the name promises, but also turns out to be faster (found this out while trying to see how much worse it would be!):
```
In [4]: a = np.hstack([np.zeros(100_000), np.random.uniform(size=100_000), np.zeros(100_000)])

In [5]: %timeit b = a.astype(bool)
288 µs ± 3.61 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [6]: %timeit b = a != 0
97.6 µs ± 185 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

[1] https://github.com/astropy/astropy/issues/10638