@BvB93 2020-08-11
I'm also seeing a noticable speedup here when running the new benchmark, 
which in and of itself would, in my opinion, make this a worthwhile change. 

I'll create a follow up pull request in a bit, as @mhvk suggested.

Before
-------
```bash
>>> python runtests.py --bench bench_trim_zeros
...
[100.00%] ··· ===================== ============ ========== ==========
              --                                   size               
              --------------------- ----------------------------------
                      dtype             3000       30000      300000  
              ===================== ============ ========== ==========
                  dtype('int64')      9.05±1μs    35.8±5μs   315±40μs 
                 dtype('float64')    15.1±0.5μs   56.1±6μs   490±90μs 
               dtype('complex128')    20.9±3μs    86.3±7μs   891±60μs 
                  dtype('bool')      10.5±0.6μs   28.4±1μs   205±5μs  
              ===================== ============ ========== ==========
```

After
-----
```bash
[100.00%] ··· ===================== ============ ========== ===========
              --                                    size               
              --------------------- -----------------------------------
                      dtype             3000       30000       300000  
              ===================== ============ ========== ===========
                  dtype('int64')     6.24±0.7μs   22.8±2μs    174±7μs  
                 dtype('float64')     11.9±1μs    40.2±4μs    321±40μs 
               dtype('complex128')    15.3±2μs    77.6±9μs   837±100μs 
                  dtype('bool')       15.0±2μs    63.6±6μs    537±30μs 
              ===================== ============ ========== ===========
```